Move IterationInterruptedException to iteratorsImpl
