+++ server/tserver/src/main/java/org/apache/accumulo/tserver/tablet/Scanner.java
@@ -27,8 +27,8 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import org.apache.accumulo.core.data.Key;
 import org.apache.accumulo.core.data.Range;
 import org.apache.accumulo.core.data.Value;
-import org.apache.accumulo.core.iterators.IterationInterruptedException;
 import org.apache.accumulo.core.iterators.SortedKeyValueIterator;
+import org.apache.accumulo.core.iteratorsImpl.system.IterationInterruptedException;
 import org.apache.accumulo.core.iteratorsImpl.system.SourceSwitchingIterator;
 import org.apache.accumulo.core.util.ShutdownUtil;
 import org.apache.accumulo.tserver.scan.ScanParameters;
